The Donation request based on New pending and completed has started showing.

+++ FoodForLife.Model/clsDonationDetails.cs
@@ -9,7 +9,7 @@ namespace FoodForLife.Model
     public class clsDonationDetails
     {
 
-        public int DonorId { get; set; }
+        public long DonorId { get; set; }
         public string Name { get; set; }
         public string ContactNumber { get; set; }
         public string EmailId { get; set; }
